Move tests from test_ethernet_servo to its parent test_servo
